Refactor Sidebar to use TypeScript
Replace EXPAND_MENU const with true TypeScript enum.
