+++ src/components/Sidebar/StickySidebar.js
@@ -15,7 +15,7 @@ const StickySidebar = styled.aside`
   height: calc((100vh - 70px) - ${NAV_HEIGHT_DESKTOP + DISTANCE_FROM_TOP_DESKTOP}px);
   overflow-y: auto;
   max-width: ${spacing.page.rightCol};
-  top: ${({ languages }) =>
+  top: ${({ 'data-languages': languages }) =>
     NAV_HEIGHT_DESKTOP + (languages ? DISTANCE_FROM_TOP_WITH_LANGUAGE_MENU_DESKTOP : DISTANCE_FROM_TOP_DESKTOP)}px;
 
   ${mq.minWidth.medium} {
