+++ src/components/Sidebar/expand-menu-enum.ts
@@ -1,0 +1,7 @@
+export enum EXPAND_MENU {
+  EXPANDED,
+  COLLAPSED,
+  SECTION_MATCH,
+  EXPAND_NEXT,
+  COLLAPSE_NEXT,
+}
