+++ src/components/Sidebar/sidebar-data.ts
@@ -1,0 +1,7 @@
+export type SidebarData = {
+  content?: SidebarData[];
+  parent?: SidebarData;
+  label: string;
+  link: string;
+  level?: number;
+};
